refactor(kolme): extract notifications websocket transport module file (#1940)
